# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project Management API",
    "version": "1.0.0",
    "description": "Mini project management API built with Golang, Gin, PostgreSQL, and JWT authentication."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "UserCreateRequest": {
        "type": "object",
        "required": ["name", "email", "password", "role"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "developer"] }
        }
      },
      "ProjectCreateRequest": {
        "type": "object",
        "required": ["name", "description"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "TaskCreateRequest": {
        "type": "object",
        "required": ["title", "description", "project_id"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "project_id": { "type": "integer" },
          "assigned_to": { "type": "integer" },
          "due_date": { "type": "string", "format": "date" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT token returned" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "List users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Paginated users" }
        }
      },
      "post": {
        "summary": "Create user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Paginated projects" }
        }
      },
      "post": {
        "summary": "Create project",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Project created" }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "summary": "List tasks with filters",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "project_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "assigned_to", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Paginated tasks" }
        }
      },
      "post": {
        "summary": "Create task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" }
        }
      }
    }
  }
}